+++ sakura_core/KeyWordSetMgr.h
@@ -87,8 +87,8 @@ public:
 	void SortKeyword(int);						// ｎ番目のセットのキーワードをソートする			//MIK
 
 	// From Here 2004.07.29 Moca 追加 可変長記憶
-	int SetKeywordArr(int, size_t, const wchar_t*);			// iniからキーワードを設定する
-	int SetKeywordArr(						// キーワードの配列から設定する
+	size_t SetKeywordArr(int, size_t, const wchar_t*);			// iniからキーワードを設定する
+	size_t SetKeywordArr(						// キーワードの配列から設定する
 		int				nIdx,				// [in] キーワードセット番号
 		size_t			nSize,				// [in] ppszKeywordArrの要素数
 		const wchar_t*	ppszKeywordArr[]	// [in] キーワードの配列(重複・長さ制限等、考慮済みであること)
