+++ sakura_core/recent/MRUFile.h
@@ -57,10 +57,10 @@ public:
 	std::vector<LPCTSTR> GetPathList() const;
 
 	//	アクセス関数
-	int Length(void) const;	//	アイテムの数。
-	int MenuLength(void) const { return t_min(Length(), recentFile.GetViewCount()); }	//	メニューに表示されるアイテムの数
+	size_t Length(void) const;	//	アイテムの数。
+	size_t MenuLength(void) const { return t_min(Length(), recentFile.GetViewCount()); }	//	メニューに表示されるアイテムの数
 	void ClearAll(void);	//	アイテムを削除～。
-	bool GetEditInfo(int num, EditInfo* pfi) const;				//	番号で指定したEditInfo（情報をまるごと）
+	bool GetEditInfo(size_t num, EditInfo* pfi) const;				//	番号で指定したEditInfo（情報をまるごと）
 	bool GetEditInfo(const TCHAR* pszPath, EditInfo* pfi) const;	//	ファイル名で指定したEditInfo（情報をまるごと）
 	void Add(EditInfo* pEditInfo);		//	*pEditInfoを追加する。
