+++ sakura_core/recent/RecentGrepFile.cpp
@@ -52,7 +52,7 @@ RecentGrepFile::RecentGrepFile()
 
 	@note	取得後のポインタはユーザ管理の構造体にキャストして参照してください。
 */
-const TCHAR* RecentGrepFile::GetItemText(int nIndex) const
+const TCHAR* RecentGrepFile::GetItemText(size_t nIndex) const
 {
 	return *GetItem(nIndex);
 }
